requirements.txt
yfinance
pandas
google-adk
streamlit
lxml
python-dotenv
google-adk
